Random chord generator
This puredata patch generates four random four-note chords and plays them in a loop. There are controls to transpose, change tempo, and output via either MIDI or sine wave oscillators. However the chords are always wrapped to fit in one octave.

The weighting of chord probabilities is adjusted to match their likelihood in a diatonic scale, with non-diatonic chords coming less likely still. The weightings are as follows:

- Minor 7: 6/18
- Major 7: 4/18
- Dominant 7: 2/18
- Half-dim 7: 2/18
- Dim 7: 2/18
- Aug 7: 1/18
- Minor #7: 1/18